requirements.txt
dotenv==0.9.9
requests==2.32.3
clickhouse-connect==0.9.2
clickhouse-driver==0.2.9
apache-airflow==2.10.2